backend/requirements.txt
flask
requests
python-dotenv
openai
langchain
pypdf
unstructured
playwright
chromadb
sentence-transformers